Reorganised test files
The test/unit subfolder wasn't really needed
