+++ test/api-views-spec.js
@@ -1,6 +1,6 @@
-import Peaks from '../../src/main';
-import WaveformOverview from '../../src/waveform-overview';
-import WaveformZoomView from '../../src/waveform-zoomview';
+import Peaks from '../src/main';
+import WaveformOverview from '../src/waveform-overview';
+import WaveformZoomView from '../src/waveform-zoomview';
 
 describe('Peaks.views', function() {
   let p;
