+++ test/segment-shape-spec.js
@@ -1,7 +1,7 @@
-import Peaks from '../../src/main';
-import SegmentShape from '../../src/segment-shape';
-import { extend } from '../../src/utils';
-import WaveformShape from '../../src/waveform-shape';
+import Peaks from '../src/main';
+import SegmentShape from '../src/segment-shape';
+import { extend } from '../src/utils';
+import WaveformShape from '../src/waveform-shape';
 
 describe('SegmentShape', function() {
   let p;
